requirements.txt
langchain-openai==0.2.6
langchain
langchain_core
langgraph
langgraph-checkpoint
langgraph-checkpoint-sqlite
langchain-community
langdetect
psycopg2-binary
tabulate
sqlalchemy
teradatasql
sqlalchemy-teradata
azure-cognitiveservices-speech
azure-cognitiveservices-search-customsearch
azure-identity
azure-search-documents==11.4.0
azure-keyvault-secrets
azure-functions
azurefunctions-extensions-bindings-blob
azure-cosmos==4.5.1
azure-mgmt-web
azure-mgmt-storage
azure-storage-blob
tiktoken
semantic-kernel==1.17.1
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
pyodbc
azure.functions
sqlparse
flask
langsmith==0.4.5
openai==1.95.1
uvicorn
pytz
requests
python-dotenv
beautifulsoup4
pyautogen==0.3.0
aiohttp==3.10.5
fuzzywuzzy
pandas
build
click
colorama
pytest
python-Levenshtein
pydantic==2.9.2
agno==1.5.5
azure-ai-inference>=1.0.0b9
gunicorn
openpyxl
ms_identity_python[flask] @ git+https://github.com/azure-samples/ms-identity-python@0.9
flask_session
cachelib

# Dependencias oficiales de Voice Live API
# aiohttp==3.11.18
azure-core==1.34.0
# azure-identity==1.22.0
certifi==2025.4.26
cffi==1.17.1
cryptography==44.0.3
numpy
pycparser==2.22
# python-dotenv==1.1.0
# requests==2.32.3
# sounddevice==0.5.1
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1

# Dependencias adicionales para el servidor Flask
# flask==3.0.0
flask-socketio==5.3.6
python-engineio==4.7.1
python-socketio==5.9.0
# asyncio-compat
